+++ spine-ue/Plugins/SpinePlugin/Source/SpinePlugin/Public/SpineAtlasAsset.h
@@ -64,12 +64,18 @@ protected:
 
 public:
 	void SetAtlasFileName(const FName &AtlasFileName);
+	void UpdateAtlasFileName(const FName &AtlasFileName);
 
 protected:
 	UPROPERTY(VisibleAnywhere, Instanced, Category = ImportSettings)
 	class UAssetImportData *importData;
 
 	virtual void PostInitProperties() override;
+#if ((ENGINE_MAJOR_VERSION >= 5) && (ENGINE_MINOR_VERSION >= 4))
+	virtual void GetAssetRegistryTags(FAssetRegistryTagsContext Context) const override;
+#else
+	virtual void GetAssetRegistryTags(TArray<FAssetRegistryTag> &OutTags) const override;
+#endif
 	virtual void Serialize(FArchive &Ar) override;
 #endif
 };